resources/code/workspace/langchain/langchain_agent/pyproject.toml
[project]
name = "langchain-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.2.10",
    "langchain-openai==1.1.10",
    "langgraph==1.0.9",
    "python-dotenv==1.2.1",
]
